Split components into different modules
