+++ src/components/calculator.js
@@ -1,4 +1,4 @@
-import PropTypes from 'prop-types';
+import Grid from './grid';
 
 const Calculator = () => (
   <div>
@@ -6,47 +6,4 @@ const Calculator = () => (
   </div>
 );
 
-const calculatorSymbols = [
-  'AC',
-  '+/-',
-  '%',
-  '÷',
-  '7',
-  '8',
-  '9',
-  'x',
-  '4',
-  '5',
-  '6',
-  '-',
-  '1',
-  '2',
-  '3',
-  '+',
-  '0',
-  '.',
-  '-',
-];
-
-const Grid = () => (
-  <div className="border" id="grid">
-    <Display />
-    {calculatorSymbols.map((symbol, index) => {
-      const id = `id${index}`;
-      return <Buttons symbol={symbol} key={symbol} id={id} />;
-    })}
-  </div>
-);
-
-const Display = () => (
-  <div className="border" id="display">
-    <h2>0</h2>
-  </div>
-);
-
-const Buttons = ({ symbol, id }) => <button id={id} type="button">{symbol}</button>;
-Buttons.propTypes = {
-  symbol: PropTypes.string.isRequired,
-  id: PropTypes.number.isRequired,
-};
 export default Calculator;
