+++ src/components/display.js
@@ -1,0 +1,7 @@
+const Display = () => (
+  <div className="border" id="display">
+    <h2>0</h2>
+  </div>
+);
+
+export default Display;
