+++ src/components/symbols.js
@@ -1,0 +1,23 @@
+const calculatorSymbols = [
+  'AC',
+  '+/-',
+  '%',
+  '÷',
+  '7',
+  '8',
+  '9',
+  'x',
+  '4',
+  '5',
+  '6',
+  '-',
+  '1',
+  '2',
+  '3',
+  '+',
+  '0',
+  '.',
+  '-',
+];
+
+export default calculatorSymbols;
